+++ CHANGELOG.md
@@ -1,6 +1,18 @@
+## 2.0.1 - 2023-06-18
+
+- Added new method for create next page request
+- Added examples
+
+## 2.0.0 - 2023-04-30
+
+- Switching to a new API client
+- Refusal of rest-api
+- Added examples
+- Added make commands
+
 ## 0.3 - 2022-12-14
 
 - changed the logic of region and language parsing
 - fix support php7.4
-- add docker-compose.yaml
-- add examples
+- Added docker-compose.yaml
+- Added examples